% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{munpr}
\alias{munpr}
\title{munpr}
\format{
A data frame with 399 rows and 4 variables:
\describe{
\item{\code{CD_MUN}}{character Nome do município}
\item{\code{NM_MUN}}{character Código do município}
\item{\code{AREA_KM2}}{double Área do município}
\item{\code{SIGLA_UF}}{double Sigla da Unidade da Federação do município}
\item{\code{geometry}}{list Geometry do município}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
munpr
}
\description{
Dados dos municípios paranaenses
}
\keyword{datasets}
